web/requirements.txt
paho-mqtt>=2.0
flask>=3.0
flask-socketio>=5.3
